refactor(#112) :: 도메인 및 service 클래스 네이밍 변경
